% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install.R
\name{available.Designs}
\alias{available.Designs}
\title{List available designs from Funz GitHub repository}
\usage{
available.Designs(refresh_repo = F)
}
\arguments{
\item{refresh_repo}{should we force refreshing GitHub Funz repositories content ?}
}
\value{
array of available designs
}
\description{
List available designs from Funz GitHub repository
}
\examples{
\dontrun{
available.Designs()
}
}
