% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install.R
\name{available.Models}
\alias{available.Models}
\title{List available models from Funz GitHub repository}
\usage{
available.Models(refresh_repo = F)
}
\arguments{
\item{refresh_repo}{should we force refreshing GitHub Funz repositories content ?}
}
\value{
array of available models
}
\description{
List available models from Funz GitHub repository
}
\examples{
\dontrun{
available.Models()
}
}
