% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/client.R
\name{CompileInput}
\alias{CompileInput}
\title{Convenience method to compile variables in parametrized file.}
\usage{
CompileInput(model, input.files, input.values, output.dir = ".")
}
\arguments{
\item{model}{name of the code wrapper to use. See .Funz.Models global var for a list of possible values.}

\item{input.files}{files to give as input for the code.}

\item{input.values}{list of variable values to compile.}

\item{output.dir}{directory where to put compiled files.}
}
\description{
Convenience method to compile variables in parametrized file.
}
\examples{
\dontrun{
CompileInput(model = "R",
             input.files = file.path(Funz:::FUNZ_HOME,"samples","branin.R"),
             input.values = list(x1=0.5, x2=0.6))
CompileInput(model = "R",
             input.files = file.path(Funz:::FUNZ_HOME,"samples","branin.R"),
             input.values = list(x1=c(0.5,.55), b=c(0.6,.7)))
}
}
