% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install.R
\name{install.Design}
\alias{install.Design}
\title{Install Funz design from local file or GitHub central repository}
\usage{
install.Design(design, force = F)
}
\arguments{
\item{design}{design to install.}

\item{force}{if already installed, reinstall.}
}
\description{
Install Funz design from local file or GitHub central repository
}
\examples{
\dontrun{
install.Design('GradientDescent')
}
}
