% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install.R
\name{install.Model}
\alias{install.Model}
\title{Install Funz model from local file or GitHub central repository}
\usage{
install.Model(model, force = F, edit.script = FALSE)
}
\arguments{
\item{model}{model to install.}

\item{force}{if already installed, reinstall.}

\item{edit.script}{open installed script for customization.}
}
\description{
Install Funz model from local file or GitHub central repository
}
\examples{
\dontrun{
install.Model('Modelica')
}
}
