% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install.R
\name{install_file.Design}
\alias{install_file.Design}
\title{Install Funz design plugin from local zip file.}
\usage{
install_file.Design(
  design.zip,
  design = gsub(".zip(.*)", "", gsub("(.*)algorithm-", "", design.zip)),
  force = F,
  ...
)
}
\arguments{
\item{design.zip}{zip file of algorithm. Usually algorithm-XYZ.zip}

\item{design}{design name (parsed in design.zip if not provided)}

\item{force}{if already installed, reinstall.}

\item{...}{optional parameters to pass to unzip()}
}
\description{
Install Funz design plugin from local zip file.
}
\examples{
\dontrun{
install_file.Design('algorithm-GradientDescent.zip')
}
}
