% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install.R
\name{install_file.Model}
\alias{install_file.Model}
\title{Install Funz model plugin from local zip file.}
\usage{
install_file.Model(
  model.zip,
  model = gsub(".zip(.*)", "", gsub("(.*)plugin-", "", model.zip)),
  force = F,
  edit.script = FALSE,
  ...
)
}
\arguments{
\item{model.zip}{zip file of plugin. Usually plugin-XYZ.zip}

\item{model}{model name (parsed in model.zip if not provided)}

\item{force}{if already installed, reinstall.}

\item{edit.script}{open installed script for customization.}

\item{...}{optional parameters to pass to unzip()}
}
\description{
Install Funz model plugin from local zip file.
}
