% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install.R
\name{install_github.Design}
\alias{install_github.Design}
\title{Install Funz design plugin from central GitHub repository.}
\usage{
install_github.Design(design, force = F)
}
\arguments{
\item{design}{design to install.}

\item{force}{if already installed, reinstall.}
}
\description{
Install Funz design plugin from central GitHub repository.
}
\examples{
\dontrun{
install_github.Design('GradientDescent')
}
}
