% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install.R
\name{install_github.Model}
\alias{install_github.Model}
\title{Install Funz model plugin from central GitHub repository.}
\usage{
install_github.Model(model, force = F, edit.script = FALSE)
}
\arguments{
\item{model}{model to install.}

\item{force}{if already installed, reinstall.}

\item{edit.script}{open installed script for customization.}
}
\description{
Install Funz model plugin from central GitHub repository.
}
\examples{
\dontrun{
install_github.Model('Modelica')
}
}
