% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install.R
\name{installed.Designs}
\alias{installed.Designs}
\title{Get available Funz Designs}
\usage{
installed.Designs()
}
\value{
array of available Designs in Funz environment.
}
\description{
Get available Funz Designs
}
\examples{
installed.Designs()
}
