% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install.R
\name{installed.Models}
\alias{installed.Models}
\title{Get available Funz Models}
\usage{
installed.Models()
}
\value{
array of available Models in Funz environment.
}
\description{
Get available Funz Models
}
\examples{
installed.Models()
}
