% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/client.R
\name{ReadOutput}
\alias{ReadOutput}
\title{Convenience method to find variables & related info. in parametrized file.}
\usage{
ReadOutput(model, input.files, output.dir, out.filter = NULL)
}
\arguments{
\item{model}{name of the code wrapper to use. See .Funz.Models global var for a list of possible values.}

\item{input.files}{files given as input for the code.}

\item{output.dir}{directory where calculated files are.}

\item{out.filter}{what output(s) to retreive in returned object.
'}
}
\value{
list of outputs & their value
}
\description{
Convenience method to find variables & related info. in parametrized file.
}
\examples{
\dontrun{
ReadOutput(model = "R", input.files = "branin.R",output.dir=".")
}
}
