% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install.R
\name{setup.Calculator}
\alias{setup.Calculator}
\title{Setup calculators in calculator.xml file.}
\usage{
setup.Calculator()
}
\description{
Setup calculators in calculator.xml file.
}
\examples{
\dontrun{
setup.Calculator()
}
}
