% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install.R
\name{setup.Model}
\alias{setup.Model}
\title{Configure model calculation entry point}
\usage{
setup.Model(model, edit.script = FALSE)
}
\arguments{
\item{model}{Name of model corresponding to given script}
}
\description{
Configure model calculation entry point
}
