% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculator.R
\name{startCalculators}
\alias{startCalculators}
\title{Start calculator instances (also named as "funz daemons")}
\usage{
startCalculators(n = 1, stdout = NULL, stderr = NULL)
}
\arguments{
\item{n}{number of calculators to start}

\item{stdout}{calculators output stream: NULL (default) or "|"}

\item{stderr}{calculators error stream: NULL (default) or "|"}
}
\value{
processx objects of started calculators
}
\description{
Start calculator instances (also named as "funz daemons")
}
\examples{
\dontrun{
# This will start 5 instances of calculator waiting for a "Run()" call
startCalculators(5)
}
}
