% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculator.R
\name{stopCalculators}
\alias{stopCalculators}
\title{Shutdown calculator instances (also named as "funz daemons")}
\usage{
stopCalculators(px)
}
\arguments{
\item{px}{array of processx objects to stop (returned values of startCalculators())}
}
\description{
Shutdown calculator instances (also named as "funz daemons")
}
\examples{
\dontrun{
calcs = startCalculators(5)
# ...
stopCalculators(calcs)
}
}
